Stop Using Overlappable Instances for ToField
My opinion is that the JSON implementation comes with a large footgun
since it can lead to `ToJSON`/`FromJSON` instances being used as the
default `ToField`/`FromField` instances. Maybe this was desired but the
overlappable instances also lead to confusing type errors.

This patch introduces `AesonField` for use with `DerivingVia`, so users
can opt-in to using JSON serialization for their fields instead of
inadvertently using aeson.

Other changes:

* Adds more explicit imports and removes redundant imports
* Amends generic `ToRow`/`FromRow` to eagerly trigger the sum type `TypeError`
* Removes a lot of redundant language extensions